Added user-defined boundary conditions (#246)
These are special boundary conditions that are never enforced
automatically, e.g., when an operator with such boundary conditions is
called, the ghost cells are simply not modified, similar to bc="none".
Instead, calling set_ghost_cells with `args={"user": value}` will allow
setting ghost points explicitly. This is an advanced feature that should
be used with care.

We also changed the default tolerance of the explicit adaptive Euler
method from 1e-4 to 1e-5 to avoid some common problems.