Fixed bug in spherical double divergence (#273)
* Added double divergence operator for tensor fields on spherical grids
* Operators on spherical symmetric grids now accept a wider range of
fields. The consistency checks on the field symmetry have been relaxed
* Scalar fields extracted from vector and tensor fields now retain
values of ghost cells
* Refactored normal boundary conditions
* Normal boundary conditions (BCs) are special BCs that only affect the
normal component of the field to which they are applied. With the new
interface, BCs need to specify all field values by default and there are
just a few special BCs that allow specifying only the normal components.
* Fixed code after rebase
* Updated documentation
* Fixed bug in double divergence of tensor fields